% Auto-generated: do not edit by hand
\name{multiFieldMapView}

\alias{multiFieldMapView}

\title{MultiFieldMapView component}

\description{
ExampleComponent is an example component. It takes a property, `label`, and displays it. It renders an input with the property `value` which is editable by the user.
}

\usage{
multiFieldMapView(id=NULL, colormap=NULL, height=NULL, highlighted=NULL,
imageHeight=NULL, imageWidth=NULL, path=NULL, polygons=NULL,
selection=NULL, width=NULL)
}

\arguments{
\item{id}{Character. The ID used to identify this component in Dash callbacks.}

\item{colormap}{Named list. }

\item{height}{Numeric. }

\item{highlighted}{Unnamed list. }

\item{imageHeight}{Numeric. }

\item{imageWidth}{Numeric. }

\item{path}{Character. }

\item{polygons}{Unnamed list. }

\item{selection}{Unnamed list. }

\item{width}{Numeric. }
}

\value{named list of JSON elements corresponding to React.js properties and their values}

